Add PropTypes checks
