+++ package.json
@@ -23,6 +23,7 @@
     "passport": "^0.4.0",
     "passport-facebook": "^2.1.1",
     "pg": "^7.4.3",
+    "prop-types": "^15.6.1",
     "react": "^16.4.0",
     "react-dom": "^16.4.0",
     "react-scripts": "1.1.4"
